用户端订单管理和百度定位距离计算是否超出配送距离

+++ sky-server/src/main/java/com/sky/mapper/OrderDetailMapper.java
@@ -11,5 +11,5 @@ public interface OrderDetailMapper {
 
     void insertBatch(List<OrderDetail> orderDetails);
     @Select("select * from order_detail where order_id = #{id}")
-    List<OrderDetail> selete(Integer id);
+    List<OrderDetail> selete(Long id);
 }
